+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file, in reverse 
 ### Changed
 - Classname changed from ServerTiming to ServerTimingMiddleware.
 - PSR-7 double pass is now supported via [tuupola/callable-handler](https://github.com/tuupola/callable-handler) library.
+- Atleast PHP 7.1 is now required.
 
 ### Removed
 -  PSR-15 is now PHP 7.x only. Support for PHP 5.X was removed.