+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## [0.6.0-dev](https://github.com/tuupola/server-timing-middleware/compare/0.5.0...0.6.0) - unreleased
+## [0.6.0](https://github.com/tuupola/server-timing-middleware/compare/0.5.0...0.6.0) - 2017-12-27
 ### Added
 - Support for the [latest version of PSR-15](https://github.com/http-interop/http-server-middleware).
 - Possibility to rename or disable default timings via options array.